+++ src/main/java/blusunrize/immersiveengineering/common/util/compat/jei/cokeoven/CokeOvenRecipeCategory.java
@@ -15,8 +15,10 @@ import blusunrize.immersiveengineering.common.blocks.IEBlocks;
 import blusunrize.immersiveengineering.common.util.compat.jei.IERecipeCategory;
 import blusunrize.immersiveengineering.common.util.compat.jei.JEIHelper;
 import blusunrize.immersiveengineering.common.util.compat.jei.JEIIngredientStackListBuilder;
+import com.mojang.blaze3d.vertex.PoseStack;
 import mezz.jei.api.constants.VanillaTypes;
 import mezz.jei.api.gui.IRecipeLayout;
+import mezz.jei.api.gui.drawable.IDrawableAnimated;
 import mezz.jei.api.gui.drawable.IDrawableStatic;
 import mezz.jei.api.gui.ingredient.IGuiFluidStackGroup;
 import mezz.jei.api.gui.ingredient.IGuiItemStackGroup;
@@ -33,14 +35,16 @@ public class CokeOvenRecipeCategory extends IERecipeCategory<CokeOvenRecipe>
 {
 	public static final ResourceLocation UID = new ResourceLocation(Lib.MODID, "cokeoven");
 	private final IDrawableStatic tankOverlay;
+	private final IDrawableAnimated flame;
 
 	public CokeOvenRecipeCategory(IGuiHelper helper)
 	{
 		super(CokeOvenRecipe.class, helper, UID, "block.immersiveengineering.coke_oven");
 		ResourceLocation background = new ResourceLocation(Lib.MODID, "textures/gui/coke_oven.png");
-		setBackground(helper.createDrawable(background, 8, 16, 142, 54));
+		setBackground(helper.createDrawable(background, 26, 16, 123, 55));
 		setIcon(new ItemStack(IEBlocks.Multiblocks.cokeOven));
 		tankOverlay = helper.createDrawable(background, 178, 33, 16, 47);
+		flame = helper.drawableBuilder(background, 177, 0, 14, 14).buildAnimated(500, IDrawableAnimated.StartDirection.TOP, true);
 	}
 
 	@Override
@@ -55,19 +59,25 @@ public class CokeOvenRecipeCategory extends IERecipeCategory<CokeOvenRecipe>
 	public void setRecipe(IRecipeLayout recipeLayout, CokeOvenRecipe recipe, IIngredients ingredients)
 	{
 		IGuiItemStackGroup guiItemStacks = recipeLayout.getItemStacks();
-		guiItemStacks.init(0, true, 21, 18);
+		guiItemStacks.init(0, true, 3, 18);
 		guiItemStacks.set(0, Arrays.asList(recipe.input.getMatchingStacks()));
 
-		guiItemStacks.init(1, false, 76, 18);
+		guiItemStacks.init(1, false, 58, 18);
 		if(!recipe.output.isEmpty())
 			guiItemStacks.set(1, recipe.output);
 
 		if(recipe.creosoteOutput > 0)
 		{
 			IGuiFluidStackGroup guiFluidStacks = recipeLayout.getFluidStacks();
-			guiFluidStacks.init(0, false, 121, 4, 16, 47, 5*FluidAttributes.BUCKET_VOLUME, false, tankOverlay);
+			guiFluidStacks.init(0, false, 103, 4, 16, 47, 5*FluidAttributes.BUCKET_VOLUME, false, tankOverlay);
 			guiFluidStacks.set(0, new FluidStack(IEContent.fluidCreosote, recipe.creosoteOutput));
 			guiFluidStacks.addTooltipCallback(JEIHelper.fluidTooltipCallback);
 		}
+	}
+
+	@Override
+	public void draw(CokeOvenRecipe recipe, PoseStack poseStack, double mouseX, double mouseY)
+	{
+		flame.draw(poseStack, 31, 20);
 	}
 }