+++ changelog.md
@@ -1,5 +1,6 @@
 ##### Version 1.16.5-5.0.8-pre
 - Add a new API interface that allows pressurized fluid in- and output above normal levels (BluSunrize)
+- Update JEI handlers for Coke Oven, Blast Furnace and Alloy Smelter to have some animated doodads (BluSunrize)
 - Fix flipped bounding boxes for the Lightning Rod (BluSunrize)
 - Fix texture glitches on Refinery model (BluSunrize)
 - Fix pipes and pumps transferring fluids much slower than planned, because pressurization didn't work (BluSunrize)